Fixed Bugs
Fixed the javascript bug with the form validation for the contact page. Added in the basic html for the apply page. Started to redo the index page to include fancy circle navigation at the bottom. Added in more images to use in the website. Fixed some of the css styling on the pages.